requirements.txt
Pillow==10.4.0
streamlit==1.30.0
streamlit==1.37.1